+++ clang/heap/take.c
@@ -4,7 +4,7 @@ fn Term heap_take(u64 loc) {
     if (__builtin_expect(term != 0, 1)) {
       return term;
     }
-    fprintf(stderr, "ERROR: heap_take saw zero at %llu in single-threaded mode\n", loc);
+    fprintf(stderr, "ERROR: heap_take saw zero at %llu in single-threaded mode\n", (unsigned long long)loc);
     abort();
   }
   for (;;) {
